move data source testing function
