+++ tensorboard/webapp/notification_center/_data_source/BUILD
@@ -31,6 +31,18 @@ tf_ts_library(
 )
 
 tf_ts_library(
+    name = "testing",
+    srcs = [
+        "testing.ts",
+    ],
+    deps = [
+        ":_data_source",
+        "@npm//@angular/core",
+        "@npm//rxjs",
+    ],
+)
+
+tf_ts_library(
     name = "_data_source_test",
     testonly = True,
     srcs = [
